pyproject.toml
[project]
name = "paper-alchemy"
version = "0.1.0"
description = "AI Research Assistant — answers research questions with cited papers"
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = "MIT"
dependencies = [
    # Web Framework
    "fastapi[standard]>=0.115.0",
    "uvicorn>=0.34.0",
    # Data Validation
    "pydantic>=2.11.0",
    "pydantic-settings>=2.8.0",
    # Database (async)
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    # Search
    "opensearch-py>=3.0.0",
    # HTTP (async)
    "httpx>=0.28.0",
    # Async File I/O
    "aiofiles>=24.0.0",
    # SSE Streaming
    "sse-starlette>=2.0.0",
    # PDF Parsing
    "docling>=2.43.0",
    # LLM & Agents
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    # Embeddings & Re-ranking
    "sentence-transformers>=5.0.0",
    # Caching
    "redis>=6.0.0",
    # Observability
    "langfuse>=3.0.0",
    # Evaluation
    "ragas>=0.2.0",
    # Utilities
    "python-dateutil>=2.9.0",
    "feedparser>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-dotenv>=0.5.2",
    "pytest-env>=1.1.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "pre-commit>=4.0.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "asgi-lifespan>=2.1.0",
    "polyfactory>=2.21.0",
    "testcontainers>=4.10.0",
    "anyio[trio]>=4.9.0",
    "aiosqlite>=0.22.1",
]

[tool.ruff]
line-length = 130
exclude = ["notebooks/**", ".venv/**", "_reference/**"]
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "ASYNC"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
env_files = ".env.test"

[tool.mypy]
explicit_package_bases = true
plugins = ["pydantic.mypy"]
